% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_alpha}
\alias{check_alpha}
\title{Check Argument alpha (Internal Use)}
\usage{
check_alpha(alpha)
}
\arguments{
\item{alpha}{Fraction of observations that must lie on each side of each split.}
}
\description{
Check Argument alpha (Internal Use)
}
