% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_honesty_inference}
\alias{check_honesty_inference}
\title{Check Arguments honesty, honesty.fraction and inference (Internal Use)}
\usage{
check_honesty_inference(honesty, honesty.fraction, inference)
}
\arguments{
\item{honesty}{Whether to grow honest forests.}

\item{honesty.fraction}{Fraction of honest sample.}

\item{inference}{Whether to conduct weight-based inference.}
}
\description{
Check Arguments honesty, honesty.fraction and inference (Internal Use)
}
