% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_keepinbag}
\alias{check_keepinbag}
\title{Check Argument keep.inbag (Internal Use)}
\usage{
check_keepinbag(keep.inbag)
}
\arguments{
\item{keep.inbag}{Save how often observations are in-bag in each tree.}
}
\description{
Check Argument keep.inbag (Internal Use)
}
