% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_maxdepth}
\alias{check_maxdepth}
\title{Check Argument max.depth}
\usage{
check_maxdepth(max.depth)
}
\arguments{
\item{max.depth}{Maximal tree depth. A value of 0 corresponds to unlimited depth, 1 to "stumps" (one split per tree).}
}
\description{
Check Argument max.depth
}
\keyword{internal}
