% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_minnodesize}
\alias{check_minnodesize}
\title{Check Argument min.node.size}
\usage{
check_minnodesize(min.node.size)
}
\arguments{
\item{min.node.size}{Minimal node size.}
}
\description{
Check Argument min.node.size
}
\keyword{internal}
