% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_nthreads}
\alias{check_nthreads}
\title{Check Argument n.threads (Internal Use)}
\usage{
check_nthreads(n.threads)
}
\arguments{
\item{n.threads}{Number of threads. Default is number of CPUs available.}
}
\value{
Appropriate value of \code{n.threads}.
}
\description{
Check Argument n.threads (Internal Use)
}
