% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ntrees}
\alias{check_ntrees}
\title{Check Argument n.trees (Internal Use)}
\usage{
check_ntrees(n.trees)
}
\arguments{
\item{n.trees}{Number of trees.}
}
\description{
Check Argument n.trees (Internal Use)
}
