% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_samplefraction}
\alias{check_samplefraction}
\title{Check Argument sample.fraction}
\usage{
check_samplefraction(sample.fraction)
}
\arguments{
\item{sample.fraction}{Fraction of observations to sample.}
}
\description{
Check Argument sample.fraction
}
\keyword{internal}
