% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_seed}
\alias{check_seed}
\title{Check Argument seed (Internal Use)}
\usage{
check_seed(seed)
}
\arguments{
\item{seed}{Random seed.}
}
\value{
Appropriate value of \code{seed}.
}
\description{
Check Argument seed (Internal Use)
}
