% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_x_y}
\alias{check_x_y}
\title{Check Arguments x and y (Internal Use)}
\usage{
check_x_y(x, y)
}
\arguments{
\item{x}{Covariate matrix (no intercept).}

\item{y}{Outcome vector.}
}
\description{
Check Arguments x and y (Internal Use)
}
