% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honesty.R
\name{class_honest_split}
\alias{class_honest_split}
\title{Honest Sample Split}
\usage{
class_honest_split(data, honesty.fraction = 0.5)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} to be split. The outcome must be located in the first column.}

\item{honesty.fraction}{Fraction of honest sample.}
}
\value{
List with elements:
  \item{\code{train_sample}}{Training sample.}
  \item{\code{honest_sample}}{Honest sample.}
}
\description{
Randomly spits the sample into a training sample and an honest sample.
}
\details{
\code{class_honest_split} looks for balanced splits, i.e., splits such as all the outcome's classes are represented
in both the training and the honest sample. After 100 trials, the program throws an error.
}
\keyword{internal}
