% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forest_weights_predicted_cpp}
\alias{forest_weights_predicted_cpp}
\title{Forest Out-of-Sample Honest Weights}
\usage{
forest_weights_predicted_cpp(
  leaf_IDs_test_list,
  leaf_IDs_honest_list,
  leaf_size_honest_list,
  w
)
}
\arguments{
\item{leaf_IDs_test_list}{List of size \code{n.trees}, storing the leaf of each tree where training units fall into.}

\item{leaf_IDs_honest_list}{List of size \code{n.trees}, storing the leaf of each tree where honest units fall into.}

\item{leaf_size_honest_list}{List of size \code{n.trees}, storing the size of the leaves of each tree computed with honest units.}

\item{w}{1 if marginal effects are being computed, 0 otherwise for normal prediction.}
}
\description{
Computes forest out-of-sample honest weights for a \code{morf.forest} object relative to the m-th class.
}
\keyword{internal}
