% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{get_pal}
\alias{get_pal}
\title{Get palette}
\usage{
get_pal(palette = NULL)
}
\arguments{
\item{palette}{Character specifying wich palette to use; default is `NULL` to randomly pick a palette}
}
\value{
Color palette
}
\description{
Get palette
}
