% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honesty.R
\name{honest_fitted}
\alias{honest_fitted}
\title{Honest In-Sample Predictions (Internal Use)}
\usage{
honest_fitted(forest, train_sample, honest_sample, m)
}
\arguments{
\item{forest}{\code{morf.forest} object.}

\item{train_sample}{Training sample.}

\item{honest_sample}{Honest sample.}

\item{m}{The class for which predictions are desired.}
}
\value{
In-sample honest predictions.
}
\description{
Computes honest in-sample predictions for a \code{morf.forest} object relative to the m-th class.
}
\details{
\code{forest} must have been grown using only the training sample. \code{honest_fitted} replaces the leaf estimates 
using the outcome from the honest sample (using the prediction method of \code{\link{morf}}).
}
