% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honesty.R
\name{honest_predictions}
\alias{honest_predictions}
\title{Honest Out-of-Sample Predictions}
\usage{
honest_predictions(forest, honest_sample, test_sample, m)
}
\arguments{
\item{forest}{\code{morf.forest} object.}

\item{honest_sample}{Honest sample.}

\item{test_sample}{Training sample.}

\item{m}{The class for which predictions are desired.}
}
\value{
Out-of-sample honest predictions.
}
\description{
Computes honest out-of-sample predictions for a \code{morf.forest} object relative to the m-th class.
}
\details{
\code{honest_predictions} replaces the leaf estimates of \code{forest} using the outcome from the associated 
honest sample (using the prediction method of \code{\link{morf}}).
}
