% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{honest_predictions_cpp}
\alias{honest_predictions_cpp}
\title{Honest Out-of-Sample Predictions}
\usage{
honest_predictions_cpp(
  unique_leaves_honest,
  y_m,
  y_m_1,
  honest_leaves,
  test_leaves
)
}
\arguments{
\item{unique_leaves_honest}{List of size \code{n.trees}, storing the unique leaf ids of each tree relative to the honest sample.}

\item{y_m}{Indicator variable, equal to 1 if the \code{y} is lower or equal than the m-th class and zero otherwise.}

\item{y_m_1}{Indicator variable, equal to 1 if the \code{y} is lower or equal than the (m-1)-th class and zero otherwise.}

\item{honest_leaves}{Matrix of size (\code{n.samples} x \code{n.trees}). The i-th row stores the id of the leaf where the i-th honest observation falls in each tree.}

\item{test_leaves}{Matrix of size (\code{n.samples} x \code{n.trees}). The i-th row stores the id of the leaf where the i-th test observation falls in each tree.}
}
\description{
Computes honest out-of-sample predictions for a morf.forest object relative to the desired class.
}
