% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-effects.R
\name{marginal_effects}
\alias{marginal_effects}
\title{Marginal Effects for Modified Ordered Random Forests}
\usage{
marginal_effects(object, data = NULL, eval = "mean", bandwitdh = 0.1)
}
\arguments{
\item{object}{\code{morf} object.}

\item{data}{Data set of class \code{data.frame} to estimate marginal effects. It must contain at least the same covariates used to train the forests. If \code{data} is \code{NULL}, marginal effects are estimated on the training sample.}

\item{eval}{Evaluation point for marginal effects. Either \code{"mean"}, \code{"atmean"} or \code{"atmedian"}.}

\item{bandwitdh}{How many standard deviations \code{x_up} and \code{x_down} differ from \code{x}.}
}
\value{
add
}
\description{
Estimates the marginal effects for a \code{morf} object.
}
\details{
\code{marginal_effects} provides a non-parametric estimation of the marginal effects for a modified ordered random
forest.\cr

Covariates with unique values between 2 and 10 are assumed to be categorical.
}
\seealso{
\code{\link{morf}}.
}
\author{
Riccardo Di Francesco
}
