% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation-metrics.R
\name{mean_ranked_score}
\alias{mean_ranked_score}
\title{Mean Ranked Probability Score}
\usage{
mean_ranked_score(y, predictions, use.true = FALSE)
}
\arguments{
\item{y}{Either the observed outcome vector or a matrix of true probabilities.}

\item{predictions}{Matrix of predictions.}

\item{use.true}{If \code{TRUE}, then the program treats \code{y} as a matrix of true probabilities.}
}
\value{
The mean ranked probability score of the method.
}
\description{
Computes the mean ranked probability score for evaluating the accuracy of ordered probability predictions.
}
\details{
If \code{use.true = FALSE}, the mean ranked probability score is computed as follows:

\deqn{\frac{1}{n} \sum_{i = 1}^n \frac{1}{M - 1} \sum_{m = 1}^M (1 (Y_i \leq m) - \hat{p}_m^* (x))^2}

otherwise:

\deqn{\frac{1}{n} \sum_{i = 1}^n \frac{1}{M - 1} \sum_{m = 1}^M (p_m^* (x) - \hat{p}_m^* (x))^2}

where:

\deqn{p_m^* (x) = P(Y_i \leq m | X_i = x)}

The second formula is useful for simulation studies.
}
\seealso{
\code{\link{morf}}, \code{\link{mean_ranked_score}}
}
\author{
Riccardo Di Francesco
}
