% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{plot.morf}
\alias{plot.morf}
\title{Plot Method for Morf Objects}
\usage{
\method{plot}{morf}(x, multiple.panels = FALSE, ...)
}
\arguments{
\item{x}{\code{morf} object.}

\item{multiple.panels}{Whether to plot each class in a separate panel.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plots a \code{morf} object.
}
\seealso{
\code{\link{morf}}, \code{\link{marginal_effects}}
}
\author{
Riccardo Di Francesco
}
