% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{print.morf}
\alias{print.morf}
\title{Print Method for morf Objects}
\usage{
\method{print}{morf}(x, ...)
}
\arguments{
\item{x}{An \code{\link{morf}} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Prints an \code{\link{morf}} object.
}
\seealso{
\code{\link{morf}}
}
\author{
Riccardo Di Francesco
}
