% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honesty.R
\name{honest_predictions}
\alias{honest_predictions}
\title{Honest Out-of-Sample Predictions}
\usage{
honest_predictions(
  forest,
  honest_sample,
  test_sample,
  y_m_honest,
  y_m_1_honest
)
}
\arguments{
\item{forest}{\code{ocf.forest} object.}

\item{honest_sample}{Honest sample.}

\item{test_sample}{Test sample.}

\item{y_m_honest}{Indicator variable, whether the outcome is smaller than or equal to the m-th class.}

\item{y_m_1_honest}{Indicator variable, whether the outcome is smaller than or equal to the (m-1)-th class.}
}
\value{
Out-of-sample honest predictions.
}
\description{
Computes honest out-of-sample predictions for an \code{ocf.forest} object.
}
\details{
\code{honest_predictions} replaces the leaf estimates of \code{forest} using the outcome from the associated 
honest sample (using the prediction method of \code{\link{ocf}}). The honest sample must not have been used
to build the trees.
}
\keyword{internal}
