% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-effects.R
\name{marginal_effects}
\alias{marginal_effects}
\title{Marginal Effects for Modified Ordered Random Forests}
\usage{
marginal_effects(
  object,
  data = NULL,
  eval = "atmean",
  bandwitdh = 0.001,
  inference = FALSE
)
}
\arguments{
\item{object}{\code{morf} object.}

\item{data}{Data set of class \code{data.frame} to estimate marginal effects. It must contain at least the same covariates used to train the forests. If \code{NULL}, marginal effects are estimated on \code{object$full_data}.}

\item{eval}{Evaluation point for marginal effects. Either \code{"mean"}, \code{"atmean"} or \code{"atmedian"}.}

\item{bandwitdh}{How many standard deviations \code{x_up} and \code{x_down} differ from \code{x}.}

\item{inference}{Whether to conduct weight-based inference. It considerably slows down the program. Not advisable if \code{eval = "mean"}.}
}
\value{
Object of class \code{morf.marginal} with elements:
  \item{\code{evaluation.type}}{Where the marginal effects are evaluated.}
  \item{\code{bandwitdh}}{The bandwitdh parameter.}
  \item{\code{n.classes}}{Number of classes.}
  \item{\code{n.samples}}{Number of samples.}
  \item{\code{n.trees}}{Number of trees in each forest.}
  \item{\code{marginal.effects}}{Matrix of marginal effects.}
  \item{\code{standard.errors}}{Matrix of standard errors, one for each marginal effect.}
  \item{\code{p.values}}{Matrix of p-values testing the null hypotheses that the marginal effect is zero.}
  \item{\code{ci.upper}}{Matrix of upper intervals.}
  \item{\code{ci.lower}}{Matrix of lower intervals.}
}
\description{
Nonparametric estimation of marginal effects using a \code{morf} object.
}
\details{
The marginal effect of the j-th covariate on \eqn{p_m \left( \cdot \right)} is defined according to the continuous or 
discrete nature of the covariate: 
 
\deqn{p_{m, j}^\prime \left( x \right) := \frac{\partial P \left( Y_i = m \, | \, X_i = x \right)}{\partial x_{j}}}

\deqn{p_{m, j}^\prime \left( x \right) := P \left( Y_i = m \, | \, X_i = \lceil x_j \rceil \right) - P \left( Y_i = m \, | \, X_i = \lfloor x_j \rfloor \right)}
    
where \eqn{x_j} is the j-th element of the vector \eqn{x} and \eqn{\lceil x_j \rceil} and \eqn{\lfloor x_j \rfloor}
correspond to \eqn{x} with its j-th element rounded up and down to the closest integer.\cr  
                     
If \code{eval = "mean"}, then the mean marginal effects are estimated:

\deqn{MME := \frac{1}{n} \sum_{i = 1}^n p_{m, j}^\prime \left( X_i \right)}

If \code{eval = "atmean"}, then the marginal effects evaluated at the mean covariates are estimated:

\deqn{MEM := p_{m, j}^\prime ( \bar{X}_i )}  

\code{eval = "atmedian"} is similar to \code{eval = "atmean"}, where the median of the covariates is used as 
evaluation point rather than the mean.\cr

The program assumes that covariates with more than ten unique values are continuous. Otherwise, covariates are assumed
to be categorical or binary.
}
\references{
\itemize{
  \item S Athey, J Tibshirani, S Wager (2019). Generalized random forests. The Annals of Statistics. \doi{10.1214/18-AOS1709}.
  \item Lechner, M., & Okasa, G. (2019). Random forest estimation of the ordered choice model. arXiv preprint arXiv:1907.02436. \doi{10.48550/arXiv.1907.02436}.
  \item Wright, M. N. & Ziegler, A. (2017). ranger: A fast implementation of random forests for high dimensional data in C++ and R. J Stat Softw 77:1-17. \doi{10.18637/jss.v077.i01}.
}
}
\seealso{
\code{\link{morf}}
}
\author{
Riccardo Di Francesco
}
