% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation-metrics.R
\name{mean_ranked_score}
\alias{mean_ranked_score}
\title{Mean Ranked Probability Score}
\usage{
mean_ranked_score(y, predictions, use.true = FALSE)
}
\arguments{
\item{y}{Either the observed outcome vector or a matrix of true probabilities (\code{num.samples} x \code{M}). See the parameter \code{use.true} and the details section.}

\item{predictions}{Matrix of predictions (\code{num.samples} x \code{M}).}

\item{use.true}{Logical. If \code{FALSE} (the default), then the program assumes that \code{y} stores the observed outcome vector, otherwise it treats \code{y} as a matrix of true probabilities.}
}
\value{
The mean ranked probability score of the method.
}
\description{
Computes the mean ranked probability score for evaluating the accuracy of ordered probability predictions.
}
\details{
If \code{use.true = FALSE}, the mean ranked probability score is computed as follows:

\deqn{\frac{1}{n} \sum_{i = 1}^n \frac{1}{M - 1} \sum_{m = 1}^M (1 (Y_i \leq m) - \hat{p}_m^* (x))^2}

otherwise:

\deqn{\frac{1}{n} \sum_{i = 1}^n \sum_{m = 1}^M (p_m^* (x) - \hat{p}_m^* (x))^2}

where:

\deqn{p_m^* (x) = P(Y_i \leq m | X_i = x)}

The second formula is useful for simulation studies.
}
