% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.morf.marginal}
\alias{print.morf.marginal}
\title{Print Method for Morf Marginal Effects}
\usage{
\method{print}{morf.marginal}(x, latex = FALSE, ...)
}
\arguments{
\item{x}{\code{morf.marginal} object.}

\item{latex}{If \code{TRUE}, prints a latex code for a table displaying the marginal effects.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Prints a \code{morf.marginal} object.
}
\seealso{
\code{\link{morf}} and \code{\link{marginal_effects}}.
}
\author{
Riccardo Di Francesco
}
