% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{print.morf.marginal}
\alias{print.morf.marginal}
\title{Print Method for morf.marginal Objects}
\usage{
\method{print}{morf.marginal}(x, latex = FALSE, ...)
}
\arguments{
\item{x}{An \code{morf.marginal} object.}

\item{latex}{If \code{TRUE}, prints LATEX code.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Prints an \code{morf.marginal} object.
}
\details{
Compilation of the LATEX code requires the following packages: \code{booktabs}, \code{float}, \code{adjustbox}. If
standard errors have been estimated, they are printed in parenthesis below each point estimate.
}
\seealso{
\code{\link{morf}}, \code{\link{marginal_effects}}.
}
\author{
Riccardo Di Francesco
}
