% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.morf}
\alias{print.morf}
\title{Print Method for Morf Objects}
\usage{
\method{print}{morf}(x, ...)
}
\arguments{
\item{x}{\code{morf} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Prints a \code{morf} object.
}
\seealso{
\code{\link{morf}}
}
\author{
Riccardo Di Francesco
}
