% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{summary.morf}
\alias{summary.morf}
\title{Summary Method for morf Objects}
\usage{
\method{summary}{morf}(object, ...)
}
\arguments{
\item{object}{An \code{\link{morf}} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summarizes an \code{\link{morf}} object.
}
\seealso{
\code{\link{morf}}, \code{\link{marginal_effects}}
}
\author{
Riccardo Di Francesco
}
