% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{summary.morf}
\alias{summary.morf}
\title{Summary Method for Morf Objects}
\usage{
\method{summary}{morf}(object, ...)
}
\arguments{
\item{object}{\code{morf} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summarizes a \code{morf} object.
}
\seealso{
\code{\link{morf}}, \code{\link{marginal_effects}}
}
\author{
Riccardo Di Francesco
}
