% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{summary.morf}
\alias{summary.morf}
\title{Summary Method for morf Objects}
\usage{
\method{summary}{morf}(object, ...)
}
\arguments{
\item{object}{An \code{\link{morf}} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summarizes an \code{\link{morf}} object.
}
\examples{
\donttest{
## Load data from orf package.
set.seed(1986)

library(orf)
data(odata)

y <- as.numeric(odata[, 1])
X <- as.matrix(odata[, -1])

## Fit morf on training sample.
forests <- morf(y, X)

## Summary.
summary(forests)}

}
\seealso{
\code{\link{morf}}, \code{\link{marginal_effects}}
}
\author{
Riccardo Di Francesco
}
