% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-s3.R
\name{summary.ocf}
\alias{summary.ocf}
\title{Summary Method for ocf Objects}
\usage{
\method{summary}{ocf}(object, ...)
}
\arguments{
\item{object}{An \code{\link{ocf}} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Summarizes an \code{\link{ocf}} object.
}
\description{
Summarizes an \code{\link{ocf}} object.
}
\examples{
\donttest{## Generate synthetic data.
set.seed(1986)

data <- generate_ordered_data(100)
sample <- data$sample
Y <- sample$Y
X <- sample[, -1]

## Fit ocf.
forests <- ocf(Y, X)

## Summary.
summary(forests)}

}
\references{
\itemize{
  \item Di Francesco, R. (2025). Ordered Correlation Forest. Econometric Reviews, 1–17. \doi{10.1080/07474938.2024.2429596}.
}
}
\seealso{
\code{\link{ocf}}, \code{\link{marginal_effects}}
}
\author{
Riccardo Di Francesco
}
