% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tree_info}
\alias{tree_info}
\title{Tree Information in Readable Format}
\usage{
tree_info(object, tree = 1)
}
\arguments{
\item{object}{\code{morf.forest} object.}

\item{tree}{Number of the tree of interest.}
}
\value{
A \code{data.frame} with the following columns:
   \item{\code{nodeID}}{Node IDs.} 
   \item{\code{leftChild}}{IDs of the left child node.} 
   \item{\code{rightChild}}{IDs of the right child node.} 
   \item{\code{splitvarID}}{IDs of the splitting variable.}
   \item{\code{splitvarName}}{Name of the splitting variable.}
   \item{\code{splitval}}{Splitting value.} 
   \item{\code{terminal}}{Logical, TRUE for terminal nodes.} 
   \item{\code{prediction}}{One column with the predicted conditional class probabilities.}
}
\description{
Extracts tree information from a \code{morf.forest} object.
}
\details{
Nodes and variables IDs are 0-indexed, i.e., node 0 is the root node. \cr

All values smaller than or equal to \code{splitval} go to the left and all values larger go to the right.
}
\examples{
## Load data from orf package.
set.seed(1986)

library(orf)
data(odata)

y <- as.numeric(odata[, 1])
X <- as.matrix(odata[, -1])

## Fit morf on training sample.
forests <- morf(y, X)

## Extract information from tenth tree of first forest.
info <- tree_info(forests$forests.info$forest.1, tree = 10)
head(info)
  
}
\seealso{
\code{\link{morf}}
}
\author{
Riccardo Di Francesco
}
