% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-info.R
\name{tree_info}
\alias{tree_info}
\title{Tree Information in Readable Format}
\usage{
tree_info(object, tree = 1)
}
\arguments{
\item{object}{\code{morf.forest} object.}

\item{tree}{Number of the tree of interest.}
}
\value{
A \code{data.frame} with the following columns:
   \item{\code{nodeID}}{The node's ID.} 
   \item{\code{leftChild}}{ID of the left child node.} 
   \item{\code{rightChild}}{ID of the right child node.} 
   \item{\code{splitvarID}}{ID of the splitting variable.}
   \item{\code{splitvarName}}{Name of the splitting variable.}
   \item{\code{splitval}}{The splitting value.} 
   \item{\code{terminal}}{Logical, TRUE for terminal nodes.} 
   \item{\code{prediction}}{One column with the predicted conditional class probabilities.}
}
\description{
Extracts tree information from a \code{morf.forest} object.
}
\details{
Nodes' and variables' IDs are 0-indexed, i.e., node 0 is the root node. \cr

All values smaller or equal than \code{splitval} go to the left and all values larger go to the right.
}
\seealso{
\code{\link{morf}}
}
\author{
Riccardo Di Francesco
}
